Organize dimension creation functions
